requirements.txt
flask
openai-whisper
torch
werkzeug
